+++ modelopt/torch/quantization/model_calib.py
@@ -1333,109 +1333,6 @@ def blockwise_weight_update(module, h, block_size, percdamp):
     module.weight.data = quantized_weight.reshape(module.weight.shape).to(module.weight.data.dtype)
 
 
-class LayerActivationGettr:
-    """Helper class for collecting layer activations during forward passes.
-
-    This class allows for sequential layer calibration by
-    patching layers to capture inputs/outputs during forward passes
-    """
-
-    def __init__(self, model: nn.Module):
-        self.model = model
-
-    @staticmethod
-    def _patch_and_initialize_layer(layer: torch.nn.Module, stop_after_collection: bool = False):
-        """Patch a layer to collect inputs and outputs during forward passes.
-
-        Args:
-            layer: The layer to patch.
-            stop_after_collection: If True, raise StopIteration after
-                collecting to enable early exit from forward pass.
-        """
-
-        def _forward_w_data_collection(self, *args, **kwargs):
-            """Custom forward that collects inputs and outputs.
-
-            Note: 'self' refers to the patched layer.
-            """
-            assert len(args) >= 1
-            self.inputs.append((args, kwargs))
-            output = self._original_forward(*args, **kwargs)
-            self.outputs.append(output)
-            if getattr(self, "_stop_after_collection", False):
-                raise StopIteration()
-            return output
-
-        bind_forward_method(layer, _forward_w_data_collection, "_original_forward")
-        layer.inputs = []
-        layer.outputs = []
-        layer._stop_after_collection = stop_after_collection
-
-    @staticmethod
-    def _unpatch_and_cleanup_layer(layer: torch.nn.Module):
-        """Restore a layer's original forward method and clean up."""
-        unpatch_forward_method(layer, "_original_forward")
-        del layer.inputs
-        del layer.outputs
-        if hasattr(layer, "_stop_after_collection"):
-            del layer._stop_after_collection
-
-    def get_input_activations(self, layer: torch.nn.Module, forward_loop: ForwardLoop) -> list:
-        """Collect input activations for a layer by running the forward loop.
-
-        Patches the layer with early stopping, runs the forward loop to collect
-        inputs from all batches. Propagation stops at the patched layer for each
-        batch (saves compute by not running deeper layers), but the forward_loop
-        continues to process all batches.
-
-        Args:
-            layer: The layer to collect inputs for.
-            forward_loop: Callable that runs calibration data through the model.
-
-        Returns:
-            List of (args, kwargs) tuples representing inputs to the layer.
-        """
-
-        # Wrap model forward to catch StopIteration per-batch
-        def _early_stop_forward(self, *args, **kwargs):
-            try:
-                return self._original_forward(*args, **kwargs)
-            except StopIteration:
-                return None  # Stop propagation but allow next batch
-
-        bind_forward_method(self.model, _early_stop_forward, "_original_forward")
-        self._patch_and_initialize_layer(layer, stop_after_collection=True)
-        try:
-            forward_loop(self.model)
-            inputs = layer.inputs.copy()
-        finally:
-            self._unpatch_and_cleanup_layer(layer)
-            unpatch_forward_method(self.model, "_original_forward")
-        return inputs
-
-    def get_output_activations(self, layer: torch.nn.Module, inputs: list) -> list:
-        """Run inputs through layer and collect outputs.
-
-        Patches the layer, forwards each input batch through it,
-        collects outputs, and unpatches the layer.
-
-        Args:
-            layer: The layer to run inputs through.
-            inputs: List of (args, kwargs) tuples - one per batch.
-
-        Returns:
-            List of outputs from the layer - one per batch.
-        """
-        self._patch_and_initialize_layer(layer, stop_after_collection=False)
-        try:
-            for args, kwargs in inputs:
-                layer(*args, **kwargs)
-            outputs = layer.outputs.copy()
-        finally:
-            self._unpatch_and_cleanup_layer(layer)
-        return outputs
-
-
 def gptq_lite(
     model: nn.Module,
     forward_loop: ForwardLoop | None = None,
